[FEATURE] Blue dot appears when clicking to select the piece to move
On mouseStop(), a blue dot is summoned to the mouse coordinates.

Also added a broken function to check for Queenside Castling. Added a
framework for reintroducing black and white alternating squares, but is
broken. To reimplement, make black_square.png not white and fix
drawHTMLChessboardLabels(). Could also be the timing.